+++ client/my-app/src/Jokes.js
@@ -1,27 +1,47 @@
-import React, {useState} from "react";
+import React, {useState, useEffect} from "react";
+
+import axios from "axios";
+
+import JokeCard from "./JokeCard"
+
+import {axiosWithAuth} from "./axiosWithAuth"
 
 
-import axiosWithAuth from "./axiosWithAuth"
 
-const Jokes = () => {
-    const displayJokes = e => {
-        e.preventDefault();
-        axiosWithAuth
-              .get("/")
-              .then(res => {
-                console.log(res)
-                        
-              })
-              .catch(err => {
-                console.log(err)
-              
-              })
-      }
+const Jokes = props => {
+
+    const [jokes, setJokes] = useState([])
+
+    const token = localStorage.getItem("token")
+
+    const options = {
+        headers: {'Authorization': token}
+      };
+
+
+
+    useEffect( () => {
+        
+    const fetchJokes = async () => {
+        
+        await axios
+           .get('http://localhost:3300/api/jokes', options)
+           .then(res => {
+                setJokes(res.data)   
+           })
+           .catch(error => {
+               console.log(error)
+           })
+       }
+       fetchJokes()
+   },[])
+
+
 
     return (
-        <>
-            
-        </>
+        <div>
+            <JokeCard jokes={jokes}/>
+        </div>
     )
 }
