+++ client/my-app/src/axiosWithAuth.js
@@ -3,9 +3,11 @@ import axios from "axios";
 export const axiosWithAuth = () => {
     const token = localStorage.getItem("token");
     return axios.create({
-        baseURL: 'http://localhistory:3300/api/jokes',
-        headers: {
-            Authorization: token
-        }
+        // baseURL: 'http://localhistory:3300/api/jokes',
+        headers: {Authorization: token}
     });
 };
+
+// const options = {
+//     headers: {'Authorization': token}
+//   };
